UART added
UART has been added to this project.
It can be used to introduce a character by keyboard and it will be represented on the display.

PS: Use TTL-3V3-RS232 USB cable. I don't know why it doesn't work when I connect the UART pins to the PSoC 5 programmer UART pins. The signal is attenuated and unreadable by a PC.